Upload python attempt at the workshop
This sketch was written on processing's python mode, and contains the results of taking an image, transforming it to its gray scale equivalent on a separate canvas, and finally, the histogram that represents the frequency of each gray value (0-255) on the image. Due to the complications of manipulating video on Python, the project will be continued on the Java mode on a separate folder.